+++ src/Makefile.test_util.include
@@ -38,6 +38,7 @@ libtest_util_a_SOURCES = \
   test/util/logging.cpp \
   test/util/mining.cpp \
   test/util/net.cpp \
+  test/util/random.cpp \
   test/util/script.cpp \
   test/util/setup_common.cpp \
   test/util/str.cpp \
