+++ src/test/denialofservice_tests.cpp
@@ -15,6 +15,7 @@
 #include <script/standard.h>
 #include <serialize.h>
 #include <test/util/net.h>
+#include <test/util/random.h>
 #include <test/util/setup_common.h>
 #include <timedata.h>
 #include <util/string.h>
