+++ src/test/miniscript_tests.cpp
@@ -6,6 +6,7 @@
 #include <string>
 #include <vector>
 
+#include <test/util/random.h>
 #include <test/util/setup_common.h>
 #include <boost/test/unit_test.hpp>
