+++ src/test/random_tests.cpp
@@ -4,6 +4,7 @@
 
 #include <random.h>
 
+#include <test/util/random.h>
 #include <test/util/setup_common.h>
 #include <util/time.h>
