+++ src/wallet/test/wallet_tests.cpp
@@ -15,6 +15,7 @@
 #include <policy/policy.h>
 #include <rpc/server.h>
 #include <test/util/logging.h>
+#include <test/util/random.h>
 #include <test/util/setup_common.h>
 #include <util/translation.h>
 #include <validation.h>
